Move test of traits to `traits` module
