+++ src/lib.rs
@@ -790,43 +790,4 @@ mod test {
         assert_abs_diff_eq!(lambert_wm1f(-1e-20), -49.962_986, epsilon = 1e-5);
         assert!(lambert_wm1f(f32::EPSILON).is_nan());
     }
-
-    #[cfg(all(feature = "traits", feature = "50bits"))]
-    #[test]
-    fn test_lambert_w0_trait_50() {
-        use super::LambertW0;
-        assert_abs_diff_eq!(
-            -2.678_794_411_714_424e-1_f64.lambert_w0(),
-            -3.993_824_525_397_807e-1
-        );
-    }
-
-    #[cfg(all(feature = "traits", feature = "24bits"))]
-    #[test]
-    fn test_lambert_w0_trait_24() {
-        use super::LambertW0;
-        assert_abs_diff_eq!(6.321_205_5e-1_f32.lambert_w0(), 4.167_04e-1, epsilon = 1e-7);
-    }
-
-    #[cfg(all(feature = "traits", feature = "50bits"))]
-    #[test]
-    fn test_lambert_wm1_trait_50() {
-        use super::LambertWm1;
-        assert_abs_diff_eq!(
-            -3.578_794_411_714_423e-1_f64.lambert_wm1(),
-            -1.253493791367214,
-            epsilon = 1e-14
-        );
-    }
-
-    #[cfg(all(feature = "traits", feature = "24bits"))]
-    #[test]
-    fn test_lambert_wm1_trait_24() {
-        use super::LambertWm1;
-        assert_abs_diff_eq!(
-            -3.578_794_3e-1_f32.lambert_wm1(),
-            -1.253_493_8,
-            epsilon = 1e-6
-        );
-    }
 }
